refactor(JOE-942): extract work-store alerts, channel-bindings, event-append
Progressive god-module split (JOE-919/JOE-942):
- work-store/alerts.ts — alert list/upsert/status domain
- work-store/channel-bindings.ts — channel bindings + claim codes
- work-store/event-append.ts — leaf appendWorkEventRow (breaks cycles)

Façade drops 5382 → 4866 LOC; fileLocBudgets ratcheted to 4900 toward 2000.
Gateway tests 1345 pass; module boundary acyclic.
